[mod/#48] onListCardClick 카드 id 적용

+++ app/src/main/java/com/haphap/app/presentation/home/HomeScreen.kt
@@ -55,7 +55,7 @@ fun HomeRoute(
 private fun HomeScreen(
     uiState: HomeContract.State,
     onRecentCardClick: (Int) -> Unit,
-    onListCardClick: () -> Unit,
+    onListCardClick: (Int) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     Box(
